Kill the assessment_options table that has once again risen from the dead

+++ db/migrate/20140430154700_drop_assessment_options_table.rb
@@ -1,5 +1,5 @@
 class DropAssessmentOptionsTable < ActiveRecord::Migration
   def change
-    drop_table :assessment_options
+    drop_table :assessment_options if table_exists? :assessment_options
   end
 end
